lombok사용전

+++ src/main/java/com/example/tanpo/repository/ExampleRepository.java
@@ -3,5 +3,6 @@ package com.example.tanpo.repository;
 import com.example.tanpo.entity.Example;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExampleRepository extends JpaRepository<Example, Integer> {
+public interface ExampleRepository extends JpaRepository<Example, Long> {
 }
+
